devops-model-garden-agent/requirements.txt
mcp
fastmcp
google-cloud-aiplatform
google-cloud-logging
python-dotenv
